Move EventSource types to WinRT.Runtime
