+++ src/WinRT.Runtime/Projections/INotifyDataErrorInfo.netstandard2.0.cs
@@ -5,6 +5,7 @@ using System;
 using System.Linq;
 using System.Runtime.InteropServices;
 using WinRT;
+using WinRT.Interop;
 
 namespace ABI.System.ComponentModel
 {
